% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGridTopo.R
\name{buildGridTopo}
\alias{buildGridTopo}
\title{buildGridTopo}
\usage{
buildGridTopo(obj, ...)
}
\arguments{
\item{obj}{spatial object}

\item{...}{options according to the spatial object}
}
\description{
buildGridTopo
}
\examples{
\dontrun{
buildGridTopo(...)
}
}
\keyword{spatialManip}

