% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcentre_offset.R
\name{cellcentre_offset}
\alias{cellcentre_offset}
\title{cellcentre_offset}
\usage{
cellcentre_offset(obj, cellsize, ...)
}
\arguments{
\item{obj}{spatial object}

\item{cellsize}{cellsize}

\item{...}{options according to the spatial object}
}
\description{
cellcentre_offset
}
\examples{
\dontrun{
cellcentre_offset(...)
}
}
\keyword{spatialManip}

