% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcentre_offset.R
\name{cellcentre_range}
\alias{cellcentre_range}
\title{cellcentre_range}
\usage{
cellcentre_range(obj, cellsize, ...)
}
\arguments{
\item{obj}{spatial object}

\item{cellsize}{cellsize}

\item{...}{options according to the spatial object}
}
\description{
cellcentre_range
}
\examples{
\dontrun{
cellcentre_range(...)
}
}
\keyword{spatialManip}

