% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellDim.R
\name{cellDim}
\alias{cellDim}
\title{cellDim}
\usage{
cellDim(obj, cellsize, ...)
}
\arguments{
\item{obj}{spatial object}

\item{cellsize}{cellsize}

\item{...}{options according to the spatial object}
}
\description{
cellDim
}
\examples{
\dontrun{
cellDim(...)
}
}
\keyword{spatialManip}

