% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{construct}
\usage{
construct(type, x, proj4S, ...)
}
\arguments{
\item{type}{type of spatial object}

\item{x}{parameters to construct the a spatial object}

\item{proj4S}{pro4string}

\item{...}{options according to the spatial object}
}
\description{
construct
}
\examples{
\dontrun{
construct(...)
}
}
\keyword{spatialManip}

