% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contain.R
\name{contain}
\alias{contain}
\title{contain}
\usage{
contain(a, b, REVERSE = FALSE)
}
\arguments{
\item{a}{spatial object}

\item{b}{spatial object}

\item{REVERSE}{boolean i.e not containing}
}
\description{
contain
}
\examples{
\dontrun{
contain()
}
}
\keyword{spatialManip}

