% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\title{coords}
\usage{
coords(obj, indice = NULL)
}
\arguments{
\item{obj}{spatial object}

\item{indice}{indice of the object in case of multi-object}
}
\description{
coords
}
\examples{
\dontrun{
coords(...)
}
}
\keyword{spatialManip}

