% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{extract}
\usage{
extract(obj, indice = NULL, ...)
}
\arguments{
\item{obj}{spatial object}

\item{indice}{indice of the object in case of multi-object}

\item{...}{options according to the spatial object}
}
\description{
extract
}
\examples{
\dontrun{
extract(...)
}
}
\keyword{spatialManip}

