% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{indice_subspatial}
\alias{indice_subspatial}
\title{indice_subspatial}
\usage{
indice_subspatial(spatial, s)
}
\arguments{
\item{spatial}{spatial grid S4 object}

\item{s}{sub spatial S4 object}
}
\description{
indice_subspatial
}
\examples{
\dontrun{
indice_subspatial()
}
}
\keyword{spatialManip}

