% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{location}
\alias{location}
\title{location}
\usage{
location(obj, points)
}
\arguments{
\item{obj}{spatial object}

\item{points}{points}

\item{...}{options according to the spatial object}
}
\description{
location
}
\examples{
\dontrun{
location(...)
}
}
\keyword{spatialManip}

