% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{mat2vec}
\alias{mat2vec}
\title{mat2vec}
\usage{
mat2vec(indice_x, indice_y, Nx, Ny)
}
\arguments{
\item{indice_x}{indice_x}

\item{indice_y}{indice_y}

\item{Nx}{Nx}

\item{Ny}{Ny}
}
\description{
mat2vec
}
\examples{
\dontrun{
mat2vec()
}
}
\keyword{spatialManip}

