% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{minmax}
\usage{
minmax(obj, ...)
}
\arguments{
\item{obj}{spatial object}

\item{...}{options according to the spatial object}
}
\description{
minmax
}
\examples{
\dontrun{
minmax(...)
}
}
\keyword{spatialManip}

