% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{proj4string}
\alias{proj4string}
\title{proj4string}
\usage{
proj4string(projargs)
}
\arguments{
\item{projargs}{projargs}
}
\description{
proj4string
}
\examples{
\dontrun{
proj4string()
}
}
\keyword{spatialManip}

