% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial2grid.R
\name{spatial2grid}
\alias{spatial2grid}
\title{spatial2grid}
\usage{
spatial2grid(obj, cellsize = NULL, ...)
}
\arguments{
\item{obj}{spatial object}

\item{cellsize}{cellsize}

\item{...}{options according to the spatial object}
}
\description{
spatial2grid
}
\examples{
\dontrun{
spatial2grid(...)
}
}
\keyword{spatialManip}

