% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice.R
\name{vec2mat}
\alias{vec2mat}
\title{vec2mat}
\usage{
vec2mat(indice, Nx, Ny)
}
\arguments{
\item{indice}{indice}

\item{Nx}{Nx}

\item{Ny}{Ny}
}
\description{
vec2mat
}
\examples{
\dontrun{
vec2mat()
}
}
\keyword{spatialManip}

