% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xGridBuild.R
\name{xGridBuild}
\alias{xGridBuild}
\title{xGridBuild}
\usage{
xGridBuild(xmin, xmax, ymin, ymax, cellsize)
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{cellsize}{cellsize}
}
\description{
xGridBuild
}
\examples{
\dontrun{
xGridBuild(...)
}
}
\keyword{spatialManip}

